+++ src/FedEx/ValidationAvailabilityAndCommitmentService/ComplexType/ServiceAvailabilityReply.php
@@ -1,0 +1,88 @@
+<?php
+namespace FedEx\ValidationAvailabilityAndCommitmentService\ComplexType;
+
+use FedEx\AbstractComplexType;
+
+/**
+ * ServiceAvailabilityReply
+ *
+ * @author      Jeremy Dunn <jeremy@jsdunn.info>
+ * @package     PHP FedEx API wrapper
+ * @subpackage  Validation Availability And Commitment Service Service
+ *
+ * @property \FedEx\ValidationAvailabilityAndCommitmentService\SimpleType\NotificationSeverityType|string $HighestSeverity
+ * @property Notification[] $Notifications
+ * @property TransactionDetail $TransactionDetail
+ * @property VersionId $Version
+ * @property ServiceAvailabilityOption[] $Options
+
+ */
+class ServiceAvailabilityReply extends AbstractComplexType
+{
+    /**
+     * Name of this complex type
+     *
+     * @var string
+     */
+    protected $name = 'ServiceAvailabilityReply';
+
+    /**
+     * Set HighestSeverity
+     *
+     * @param \FedEx\ValidationAvailabilityAndCommitmentService\SimpleType\NotificationSeverityType|string $highestSeverity
+     * @return $this
+     */
+    public function setHighestSeverity($highestSeverity)
+    {
+        $this->values['HighestSeverity'] = $highestSeverity;
+        return $this;
+    }
+
+    /**
+     * Set Notifications
+     *
+     * @param Notification[] $notifications
+     * @return $this
+     */
+    public function setNotifications(array $notifications)
+    {
+        $this->values['Notifications'] = $notifications;
+        return $this;
+    }
+
+    /**
+     * Set TransactionDetail
+     *
+     * @param TransactionDetail $transactionDetail
+     * @return $this
+     */
+    public function setTransactionDetail(TransactionDetail $transactionDetail)
+    {
+        $this->values['TransactionDetail'] = $transactionDetail;
+        return $this;
+    }
+
+    /**
+     * Set Version
+     *
+     * @param VersionId $version
+     * @return $this
+     */
+    public function setVersion(VersionId $version)
+    {
+        $this->values['Version'] = $version;
+        return $this;
+    }
+
+    /**
+     * Set Options
+     *
+     * @param ServiceAvailabilityOption[] $options
+     * @return $this
+     */
+    public function setOptions(array $options)
+    {
+        $this->values['Options'] = $options;
+        return $this;
+    }
+}